refactor: sunshine data is not related to db, move it out of db/
