+++ src/pages/sunshine/[entity]/[id]/overview.tsx
@@ -49,7 +49,7 @@ import {
   fetchOperationalStandards,
   fetchOperatorCostsAndTariffsData,
   fetchPowerStability,
-} from "src/lib/db/sunshine-data";
+} from "src/lib/sunshine-data";
 import { defaultLocale } from "src/locales/config";
 
 import {
